requirements.txt
sqlalchemy
pymysql
dotenv
flask
mysql-connector-python